Add author names to main navigation in footer
Addresses feedback in #1

There might be other ways to handle this in future that involve changing
the design of the authors page but for now with 4 authors listed, should
work fine. For visual balance, it would also be good to add some extra
links to the existing main nav in the footer: research papers and
about/statement perhaps.